Adjust base price of product for price include tax
The final price of the product has been calculated correctly through
use of the shopping_cart class outcome; however, the base price
of the product(s) did not take the tax into account and therefore
needs to be added in which is done here.

In issue #14, there was a suggestion to modify this approach awaiting
reply from @webchills.  In absence of response expecting further
revision, this fixes #14.