Begin reducing DPU on screen updates
Code modifications to reduce the price flashing when selecting
attributes that do not affect price (or entering text in a text
box that does not change the price of the product).

Attempt to get product's display price only for non-zero product.
Prevent loading if the price is zero and there are no option ids.

Support preventing skipping the load of DPU for a product.

By $optionIds not being empty, if a product is identified
by `zen_get_products_price_is_free` to be true, then it is
possible to not block the load of DPU.


In order to determine if this module should be active on a given
page or for given fields, it is important to understand what
attribute information affects the price.  If there is nothing
that affects the price, then the returned array will be empty
(`empty($results) == true`), otherwise, the array will have
size and values that can be used to further hone the operations.

This is the first attempt at adding code to the jscript_ file to
try to bypass attaching an event listener to an attribute that
doesn't affect price.

Now, the function was written to account for and create a list of
option names that DO change the price; however, perhaps it would
be better to make the function identify the items that do not
affect price.  This way if the value is encountered then can
prevent updating, otherwise allow updating for all other items...

In earlier review to implement it was perceived
that there were variables that would be loaded before the
jscript_ file.  Unfortunately the sought after variables are loaded
later. This generates the desired variables which will then
also be reset later.

At this time no additional notifiers have been added to handle
this particular case.

This now works such that for the given attribute type, the
monitoring javascript is not provided unless the applicable
type possibly is included by there being an attribute
associated with the product.

If there are attributes, then the quantity box will be monitored
even if it is not present (ie. if the text field has the name
cart_quantity, then it will be processed to update.  If there
are no attributes and the quantity box is not provided then
this module will be disabled effectively as there is the
perception that nothing will modify the price.  This does also
mean that the resulting price text is not modified by DPU and
if there is some alternate text to be displayed that it must be
modified in the ZC language define(s) for that data type.

Incorporates text and file types and notifier for others

This commit adds the sought after content for the name field of the
attributes when the attribute is a text entry type or file entry type.
If it is neither of those, then there is a notifier that can be used:
NOTIFY_DYNAMIC_PRICE_UPDATER_ATTRIBUTE_ID_TEXT
to further modify the data associated.

Note, that the data structure of this effort may be changed in the
near future to add some flexibility and to reduce the content sent
to the visitor by further use of arrays.