Determine attributes that affect price
In order to determine if this module should be active on a given page or for given fields, it is important to understand what attribute information affects the price.  If there is nothing that affects the price, then the returned array will be empty (`empty($results) == true`), otherwise, the array will have size and values that can be used to further hone the operations.