Correct variable existence issue.
In earlier review for prevention of loading DPU it was perceived
that there were variables that would be loaded before the
jscript_ file.  Unfortunately the sought after variables are loaded
later. This commit generates the desired variables which will then
also be reset later.

At this time no additional notifiers have been added to handle
this particular case.