Correct javascript display logic
In initial testing, parts of this new feature appeared to work
correctly to provide update capability when it was to be applied;
however, it was found that updating also occurred when it was
desired not to provide updates.  This now corrects the issue
such that for the given attribute type, the monitoring javascript
is not provided unless the applicable type possibly is included
by there being an attribute associated with the product.

If there are attributes, then the quantity box will be monitored
even if it is not present (ie. if the text field has the name
cart_quantity, then it will be processed to update.  If there
are no attributes and the quantity box is not provided then
this module will be disabled effectively as there is the
perception that nothing will modify the price.  This does also
mean that the resulting price text is not modified by DPU and
if there is some alternate text to be displayed that it must be
modified in the ZC language define(s) for that data type.