Incorporate text and file types and notifier for others
This commit adds the sought after content for the name field of the
attributes when the attribute is a text entry type or file entry type.
If it is neither of those, then there is a notifier that can be used:
NOTIFY_DYNAMIC_PRICE_UPDATER_ATTRIBUTE_ID_TEXT
to further modify the data associated.

Note, that the data structure of this effort may be changed in the
near future to add some flexibility and to reduce the content sent
to the visitor by further use of arrays.