Revise onpage uri and logging attempts
Modified the internal URI generation so that on page load, there is no longer a direct link provided to each individual instance of executing the ajax query link.  This should suffice even for URI rewriters; however, that also depends on the rewriter and how it operates.

Will likely have to hear back from others including what URI rewriter they are using and how it works as to how to modify this further to prevent the full URI being in a single clickable link.

The goal is that the page will serve up the <base href> tag with a full address to the base of the store where ajax.php is located.  Therefore the link to execute the ajax will be properly formed by reaching to the ajax.php file off of the current base href.

Additionally added a check similar to what is used in the template jscript_framework.php file to test for the console.log before attempting to push anything to it.