Provide specific identifier for extra selections
On a somewhat rare occasion, an attribute that did not have
a selected options value would be assigned a value that may
relate to a real value and therefore cause an inaccurate
display of the product's price.  Now the value is set to
false which can be tested with an absolute comparison. The
normal case where this has been seen as necessary was for
checkbox type attributes.