Offer chance for faster processing
Adds an isset check for conditions where it is desired to determine
if a key is in the array.  The isset condition is faster, but
if the value is null then is treated as false so still doing
the check against the keys as well if the array value comes back
as non-existent.