Remove products_id from ajax request
As a result of other changes to ZC, providing only the
products_id without a main_page when calling the ajax
command results in continuous reload of the page.  The
products_id is provided in other ways and can be used
to accomplish the same task which has been to update
the who's online information to show the page on which
the customer is visiting rather than just the call to
the associated ajax parameters.

Fixes #21