Check text value against database constant.
The test for disabling version checking against the Zen Cart
site incorrectly assumed that a string value of `'false'` would
reflect as a boolean value of false in an `if` statement.
Not so.. It appears as a loose `true` because there is content
and it is not a false.

Anyways, chose to compare against not equals false so that
the available values could be modified to something else as
desired by the site owner with the expectation that if it is
`'false'` then these lines should not be executed.